lab6_code_report/requirements.txt
diffusers==0.30.0
matplotlib==3.8.4
numpy==2.1.0
Pillow==10.4.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
